[SPARK-38504][SQL] Cannot read TimestampNTZ as TimestampLTZ
